Read Triggered Compactoin in Get()
Summary: Trigger a compaction if a file is read more than a certain threshold, and the DB doesn't need to do other compactions. LevelDB supports that but we removed it for performance reason. Now we are bringing it back based on sampled stats.

Run db_bench starting with data filled with L0, L1 and L2. After running it long enough, all data is compacted to L2

Test Plan: Will add a new unit test; run several db_bench setting.